pythonProject/requirements.txt
langchain
openai==0.27.8
tiktoken
requests
pydantic
pinecone-client
